Move shared components into their own pattern library dir
In order to help our future developers, we needed a clear
indicator what components are built for reuse and which are not.
Moving shared components into a patternLibrary lib sends a clear
indication to users which components are meant for broad use. It
also serves as a place to put future components.
