+++ src/patternLibrary/ContentSection/ContentSection.js
