+++ src/patternLibrary/ContentSection/ContentSection.scss
