+++ src/patternLibrary/Icon/Icon.js
