+++ src/patternLibrary/Icon/Icon.scss
