+++ src/patternLibrary/Icon/index.js
