+++ src/patternLibrary/Layout/Layout.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Helmet from "react-helmet";
 import cx from "classnames";
-import Header from "../Header";
+import Header from "../../components/Header";
 
 import "./Layout.scss";
