+++ src/patternLibrary/Text/Text.js
