+++ src/patternLibrary/Text/index.js
